% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_analysis.R
\name{wb_analysis}
\alias{wb_analysis}
\title{wb_analysis: World Bank Data Analysis}
\usage{
wb_analysis(countries, indicators)
}
\arguments{
\item{countries}{Vector of strings. It must include the ISO Alpha-2 code of the countries of interest.}

\item{indicators}{Vector of strings. It must include the ID of the indicators from "The World Bank" datasets.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
    \item ISO-2
    \item Country
    \item Year
    \item \code{indicators}
}

It also prints plots for each indicators and, if there are more than one, the correlation matrix.
}
\description{
Returns a data frame of the chosen indicators for the selected countries; displays plots for each indicator and prints a correlation matrix in case of two or more indicators.
}
\examples{
wb_analysis(countries = "IT", indicators = c("NY.GDP.MKTP.CD", "SP.DYN.LE00.IN"))

wb_analysis(countries = c("US", "CN", "RU"), indicators = c("SP.POP.TOTL", "SL.TLF.TOTL.IN", "NY.GDP.PCAP.CD"))

}
\seealso{
\link[WDI]{WDIsearch}, for indicators' ID.
}
\author{
Sergio Picascia \email{sergio.picascia@studenti.unimi.it}
}
