% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cor.R
\name{wb_cor}
\alias{wb_cor}
\title{wb_cor: computing correlations for World Bank data}
\usage{
wb_cor(countries, indicators, dataset)
}
\arguments{
\item{countries}{Vector of strings. It must include the ISO Alpha-2 code of the countries of interest.}

\item{indicators}{Vector of strings. It must include the ID of the indicators from "The World Bank" datasets.}

\item{dataset}{Data frame. It is recommended to create one using \link{wb_tidy} in order to avoid possible errors.}
}
\description{
Prints a matrix which values are the correlations between the chosen indicators for each of the selected countries.
There are two possible usage:
\itemize{
    \item it is possible to use a local data frame, using it as parameter of \code{dataset};
    \item if ther argument \code{dataset} is missing, you must indicate the arguments \code{countries} and \code{indicators} and a dataset will be generated using \link{wb_tidy}
}
}
\examples{
wb_cor(countries = c("US", "CN", "RU"), indicators = c("SP.POP.TOTL", "SL.TLF.TOTL.IN", "NY.GDP.PCAP.CD"))

\dontrun{
wb_cor(dataset = my_data)
}

}
\seealso{
\link{wb_tidy}
}
\author{
Sergio Picascia \email{sergio.picascia@studenti.unimi.it}
}
