% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_plot.R
\name{wb_plot}
\alias{wb_plot}
\title{wb_plot: showing plots for World Bank data}
\usage{
wb_plot(countries, indicators, dataset)
}
\arguments{
\item{countries}{Vector of strings. It must include the ISO Alpha-2 code of the countries of interest.}

\item{indicators}{Vector of strings. It must include the ID of the indicators from "The World Bank" datasets.}

\item{dataset}{Data frame. It is recommended to create one using \link{wb_tidy} in order to avoid possible errors.}
}
\description{
Displays a line plot for each indicator.
There are two possible usage:
\itemize{
    \item it is possible to use a local data frame, using it as parameter of \code{dataset};
    \item if ther argument \code{dataset} is missing, you must indicate the arguments \code{countries} and \code{indicators} and a dataset will be generated using \link{wb_tidy}
}
}
\examples{
wb_plot(countries = "IT", indicators = c("NY.GDP.MKTP.CD", "SP.DYN.LE00.IN"))

\dontrun{
wb_plot(dataset = my_data)
}

}
\seealso{
\link{wb_tidy}
}
\author{
Sergio Picascia \email{sergio.picascia@studenti.unimi.it}
}
