% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_tidy.R
\name{wb_tidy}
\alias{wb_tidy}
\title{wb_tidy: tidy World Bank dataset}
\usage{
wb_tidy(countries, indicators)
}
\arguments{
\item{countries}{Vector of strings. It must include the ISO Alpha-2 code of the countries of interest.}

\item{indicators}{Vector of strings. It must include the ID of the indicators from "The World Bank" datasets.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
    \item ISO-2
    \item Country
    \item Year
    \item \code{indicators}
}
}
\description{
Returns a data frame of the chosen indicators for the selected countries.
}
\examples{
wb_tidy(countries = "IT", indicators = c("NY.GDP.MKTP.CD", "SP.DYN.LE00.IN"))

wb_tidy(countries = c("US", "CN", "RU"), indicators = c("SP.POP.TOTL", "SL.TLF.TOTL.IN", "NY.GDP.PCAP.CD"))

}
\seealso{
\link[WDI]{WDIsearch}, for indicators' ID.
}
\author{
Sergio Picascia \email{sergio.picascia@studenti.unimi.it}
}
