% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbanalysis.R
\docType{package}
\name{wbanalysis}
\alias{wbanalysis}
\alias{wbanalysis-package}
\title{wbanalysis: Analysis of "The World Bank" data}
\description{
Easy way to quickly analyze World Bank datasets and generate a tidy dataframe.
}
\author{
Sergio Picascia \email{sergio.picascia@studenti.unimi.it}
}
